requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==5.2.0
pytz==2018.7
whitenoise==4.1.2
psycopg2==2.7.6
django-rest-auth==0.9.3
